fix readjustment logic
since forever ago logo.text was NULL, turning a crucial part of the offset
calculation noop.

to preserve the intention of the code we adjust the offset Y such that it
is *always* slightly below the spinner. this should prevent the label
text from ever overlapping with the progress indication.

(this is only applied to master at the time of writing as there is no
 known bug report of the offset being wrong, so conservatively this only
 gets changed in master at the off chance of it having unintended side
 effects)